Share _USER_DATA/_ITEM_DATA structs between Ebenezer and Aujard
